Enabling work. Split Integration tests & refactor creation of concern
