+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/ConcernsRecord/MappingProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+
+namespace ConcernsCaseWork.API.Features.ConcernsRecord
+{
+	using ConcernsCaseWork.Data.Models;
+
+	public class MappingProfile : Profile
+	{
+		public MappingProfile()
+		{
+			CreateMap<ConcernsRecord, GetByID.Result>()
+				.ForMember(dest => dest.CaseUrn, opt => opt.MapFrom(src => src.ConcernsCase.Urn));
+		}
+	}
+}
